requirements.txt
flask
pywebview
static-ffmpeg
pyinstaller
